requirements.txt
Flask>=1.0.0
rfc3339
ansicolors
